Updating program
Adding functionality to generate and search for cards.

Adding ability to convert decks into codes that the actual game can use and back again.

Reworking boss mechanic to be revealed at the start of the world and apply throughout the entirety. Introducing "mook" mechanic so that every location has an enemy to fight and the location/HP systems are uniform.