Additional functionality
Created a working prototype of the in-client deck builder.

Confirmed that deck pasting works into the actual game. Still need to paste decks into my client from the game.

Updated worlds to show enemy and location in every section. Modified worlds to be created as you reach them, so that the boss can always be a random free agent that won't be spoiled.

Customized the look of boss nodes on the world map to make it clearer when they are selectable.

Added base functionality for the World Clear screen and World 9.