Updates 10/20
Adding Exhausted, Pig, and Raptor status effects. Setting cards to automatically have a chance of becoming exhausted if used multiple times in a row. Chance increases the more times a card is used.

Added a tracker to keep track of MIA cards that are sent away. Cards that are sent away now return at the beginning of the world instead of as a reward in one of the sections.

Modified HP display to be easier to change and reworked calculation to make it impossible to raise the enemy's HP above base.